Refactor out check_pfs_for_production
Also move all tests out of test_service_registry.py to
test_pathfinding.py
