+++ raiden/tests/integration/network/proxies/test_service_registry.py
@@ -1,23 +1,14 @@
 from unittest.mock import Mock, patch
 
 import pytest
-import requests
-from eth_utils import (
-    is_canonical_address,
-    is_same_address,
-    to_canonical_address,
-    to_checksum_address,
-)
+from eth_utils import to_canonical_address
 
-from raiden.constants import Environment, RoutingMode
 from raiden.exceptions import BrokenPreconditionError
-from raiden.network.pathfinding import configure_pfs_or_exit, get_random_pfs, get_valid_pfs_url
-from raiden.settings import DEFAULT_PATHFINDING_MAX_FEE
+from raiden.network.pathfinding import get_random_pfs, get_valid_pfs_url
 from raiden.tests.utils.factories import HOP1
-from raiden.tests.utils.mocks import mocked_failed_response, mocked_json_response
 from raiden.tests.utils.smartcontracts import deploy_service_registry_and_set_urls
 from raiden.utils import privatekey_to_address
-from raiden.utils.typing import ChainID, FeeAmount, TokenNetworkRegistryAddress
+from raiden.utils.typing import FeeAmount, TokenNetworkRegistryAddress
 
 token_network_registry_address_test_default = TokenNetworkRegistryAddress(
     to_canonical_address("0xB9633dd9a9a71F22C933bF121d7a22008f66B908")
@@ -81,183 +72,3 @@ def test_service_registry_random_pfs(
         assert (
             get_random_pfs(c1_service_proxy, "latest", pathfinding_max_fee=FeeAmount(100)) in urls
         )
-
-
-def test_configure_pfs(service_registry_address, private_keys, web3, contract_manager):
-    chain_id = ChainID(int(web3.net.version))
-    service_registry, urls = deploy_service_registry_and_set_urls(
-        private_keys=private_keys,
-        web3=web3,
-        contract_manager=contract_manager,
-        service_registry_address=service_registry_address,
-    )
-    json_data = {
-        "price_info": 0,
-        "network_info": {
-            "chain_id": chain_id,
-            "registry_address": to_checksum_address(token_network_registry_address_test_default),
-        },
-        "message": "This is your favorite pathfinding service",
-        "operator": "John Doe",
-        "version": "0.0.1",
-        "payment_address": to_checksum_address(privatekey_to_address(private_keys[0])),
-    }
-
-    response = mocked_json_response(response_data=json_data)
-
-    # With local routing configure_pfs should raise assertion
-    with pytest.raises(AssertionError):
-        _ = configure_pfs_or_exit(
-            environment_type=Environment.PRODUCTION,
-            pfs_url="",
-            routing_mode=RoutingMode.LOCAL,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-
-    # With private routing configure_pfs should raise assertion
-    with pytest.raises(AssertionError):
-        _ = configure_pfs_or_exit(
-            environment_type=Environment.PRODUCTION,
-            pfs_url="",
-            routing_mode=RoutingMode.PRIVATE,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-
-    # Asking for auto address
-    # To make this deterministic we need to patch the random selection function
-    patch_random = patch("raiden.network.pathfinding.get_random_pfs", return_value="http://foo")
-    with patch.object(requests, "get", return_value=response), patch_random:
-        config = configure_pfs_or_exit(
-            environment_type=Environment.PRODUCTION,
-            pfs_url="auto",
-            routing_mode=RoutingMode.PFS,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-    assert config.url in urls
-    assert is_canonical_address(config.payment_address)
-
-    # Configuring a valid given address
-    given_address = "http://foo"
-    with patch.object(requests, "get", return_value=response):
-        config = configure_pfs_or_exit(
-            environment_type=Environment.PRODUCTION,
-            pfs_url=given_address,
-            routing_mode=RoutingMode.PFS,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-    assert config.url == given_address
-    assert is_same_address(config.payment_address, json_data["payment_address"])
-    assert config.price == json_data["price_info"]
-
-    # Configuring an address that doesn't match the registered url should error
-    given_address = "http://ourgivenaddress"
-    with pytest.raises(SystemExit):
-        with patch.object(requests, "get", return_value=response):
-            configure_pfs_or_exit(
-                environment_type=Environment.PRODUCTION,
-                pfs_url=given_address,
-                routing_mode=RoutingMode.PFS,
-                service_registry=service_registry,
-                node_network_id=chain_id,
-                token_network_registry_address=token_network_registry_address_test_default,
-                pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-            )
-    # but work when in development mode
-    with patch.object(requests, "get", return_value=response):
-        config = configure_pfs_or_exit(
-            environment_type=Environment.DEVELOPMENT,
-            pfs_url=given_address,
-            routing_mode=RoutingMode.PFS,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-        assert config.url == given_address
-
-    # Configuring an pfs payment address that isn't registered should error
-    given_address = "http://foo"
-    unregistered_json_data = json_data.copy()
-    unregistered_json_data["payment_address"] = "0x2222222222222222222222222222222222222221"
-    unregistered_response = mocked_json_response(response_data=unregistered_json_data)
-    with pytest.raises(SystemExit):
-        with patch.object(requests, "get", return_value=unregistered_response):
-            configure_pfs_or_exit(
-                environment_type=Environment.PRODUCTION,
-                pfs_url=given_address,
-                routing_mode=RoutingMode.PFS,
-                service_registry=service_registry,
-                node_network_id=chain_id,
-                token_network_registry_address=token_network_registry_address_test_default,
-                pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-            )
-    # but work when in development mode
-    with patch.object(requests, "get", return_value=unregistered_response):
-        config = configure_pfs_or_exit(
-            environment_type=Environment.DEVELOPMENT,
-            pfs_url=given_address,
-            routing_mode=RoutingMode.PFS,
-            service_registry=service_registry,
-            node_network_id=chain_id,
-            token_network_registry_address=token_network_registry_address_test_default,
-            pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-        )
-        assert config.url == given_address
-
-    # Bad address, should exit the program
-    response = mocked_failed_response(error=requests.RequestException(), status_code=400)
-    bad_address = "http://badaddress"
-    with pytest.raises(SystemExit):
-        with patch.object(requests, "get", side_effect=requests.RequestException()):
-            # Configuring a given address
-            _ = configure_pfs_or_exit(
-                environment_type=Environment.PRODUCTION,
-                pfs_url=bad_address,
-                routing_mode=RoutingMode.PFS,
-                service_registry=service_registry,
-                node_network_id=chain_id,
-                token_network_registry_address=token_network_registry_address_test_default,
-                pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-            )
-
-    # Addresses of token network registries of pfs and client conflict, should exit the client
-    response = mocked_json_response(response_data=json_data)
-    with pytest.raises(SystemExit):
-        with patch.object(requests, "get", return_value=response):
-            _ = configure_pfs_or_exit(
-                environment_type=Environment.PRODUCTION,
-                pfs_url="http://foo",
-                routing_mode=RoutingMode.PFS,
-                service_registry=service_registry,
-                node_network_id=chain_id,
-                token_network_registry_address=TokenNetworkRegistryAddress(
-                    to_canonical_address("0x2222222222222222222222222222222222222221")
-                ),
-                pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-            )
-
-    # ChainIDs of pfs and client conflict, should exit the client
-    response = mocked_json_response(response_data=json_data)
-    with pytest.raises(SystemExit):
-        with patch.object(requests, "get", return_value=response):
-            configure_pfs_or_exit(
-                environment_type=Environment.PRODUCTION,
-                pfs_url="http://foo",
-                routing_mode=RoutingMode.PFS,
-                service_registry=service_registry,
-                node_network_id=ChainID(chain_id + 1),
-                token_network_registry_address=token_network_registry_address_test_default,
-                pathfinding_max_fee=DEFAULT_PATHFINDING_MAX_FEE,
-            )
